translated some parts of website. fixed queries for selecting and ordering places in irakli_routes.
working on "General error: 2014 Cannot execute queries while other unbuffered queries are active" error, which was found when i was fixing tag-connecting.
without success so far.
the function doesn't add duplicate entries anymore.
it doesn't add anything because of that error. but I'm pretty sure that if we fix that error, duplicate entries won't be added anymore.